Split AnsysSolution into subclasses
Create dedicated solution class for each `ansys_tool` alternative, and include only relevant parameters for each solution type.

Related changes:

- Introduce `get_parameters` function in `Solution` so that ClassVar parameter `ansys_tool` ends up into json "parameters".

- Generalize `find_varied_parameters` to cover variable "parameters" keys in different json files. This is necessary because keys in "parameters" depend on solution type.

- Raise error if `export_ansys_json` is going to override json file. This is useful to make sure that user remembers to set unique names for simulation and solution classes.

- Rename AnsysEigenmodeSolution parameter "frequency" to "min_frequency", because it means minimum allowed eigenmode frequency. The change is applied in 5 export scripts.

- Rename AnsysHfssSolution parameter "hfss_capacitance_export" to "capacitance_export". No usage in export scripts.

- Remove unused export parameter `sweep_enabled` from `cpw_circle_sim.py`.

- Modernize export scripts `swissmon_fluxline_sim.py` and `swissmon_epr_sim.py` (renamed) to use solution sweep feature.

- Modernize `tls_waveguide_sim.py` to use AnsysHfssSolution.

- Change renamed export parameters in user guide

+++ klayout_package/python/kqcircuits/simulations/export/ansys/ansys_export.py
@@ -23,7 +23,7 @@ import logging
 from typing import Optional, Union, Sequence, Tuple
 from pathlib import Path
 
-from kqcircuits.simulations.export.ansys.ansys_solution import AnsysSolution
+from kqcircuits.simulations.export.ansys.ansys_solution import AnsysSolution, get_ansys_solution
 from kqcircuits.simulations.export.simulation_export import (
     copy_content_into_directory,
     get_post_process_command_lines,
@@ -71,8 +71,13 @@ def export_ansys_json(simulation: Simulation, solution: AnsysSolution, path: Pat
 
     # write .json file
     json_file_path = str(path.joinpath(full_name + ".json"))
-    with open(json_file_path, "w") as fp:
-        json.dump(json_data, fp, cls=GeometryJsonEncoder, indent=4)
+    if not Path(json_file_path).exists():
+        with open(json_file_path, "w") as fp:
+            json.dump(json_data, fp, cls=GeometryJsonEncoder, indent=4)
+    else:
+        raise ValueError(
+            f"Json file '{full_name}.json' already exists. Make sure that simulations and solutions have unique names."
+        )
 
     return json_file_path
 
@@ -175,13 +180,9 @@ def export_ansys(
     write_commit_reference_file(path)
     copy_content_into_directory(ANSYS_SCRIPT_PATHS, path, script_folder)
     json_filenames = []
+    common_sol = None if all(isinstance(s, Sequence) for s in simulations) else get_ansys_solution(**solution_params)
     for sim_sol in simulations:
-        if isinstance(sim_sol, Sequence):
-            simulation, solution = sim_sol
-        else:
-            simulation = sim_sol
-            solution = AnsysSolution(**solution_params)
-
+        simulation, solution = sim_sol if isinstance(sim_sol, Sequence) else (sim_sol, common_sol)
         try:
             json_filenames.append(export_ansys_json(simulation, solution, path))
         except (IndexError, ValueError, Exception) as e:  # pylint: disable=broad-except
